Support mixed-case elements & attributes in inline SVG/MathML
Consistent with the HTML specs, this is done as a fix-up after parsing,
with an explicit list of recognized attribute and element names.
Because it's a long list, I've included it as a separate JSON file.
Values are copy-pasted straight from the tables in
https://html.spec.whatwg.org/multipage/parsing.html

In XML mode, the original case is preserved by the parser,
no fix-up required.